rearrange the project: fragments & adapters should be under "client" package, the entities under "models" package and the future dao services/EJBs under "server" package
