--- .idea/libraries/Gradle____app_libs_jtds_1_3_1_jar.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: ./app/libs/jtds-1.3.1.jar">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/app/libs/jtds-1.3.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
