+++ app/src/main/java/com/example/musimatch/client/AllPostsFragment.java
@@ -1,4 +1,4 @@
-package com.example.musimatch;
+package com.example.musimatch.client;
 import android.os.Bundle;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -6,7 +6,9 @@ import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import com.example.musimatch.adapters.PostAdapter;
+
+import com.example.musimatch.R;
+import com.example.musimatch.client.adapters.PostAdapter;
 import com.example.musimatch.models.Post;
 import com.example.musimatch.models.PostModel;
 import java.util.ArrayList;
