+++ app/src/main/res/layout/fragment_all_posts.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AllPostsFragment">
+    tools:context=".client.AllPostsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/allPostsFRG_recyclerView"
