+++ app/src/main/res/layout/fragment_edit_post.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/dark_purple"
-    tools:context=".EditPostFragment">
+    tools:context=".client.EditPostFragment">
 
     <TextView
         android:id="@+id/textView11"
