+++ app/src/main/res/layout/fragment_edit_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/dark_purple"
-    tools:context=".EditProfileFragment">
+    tools:context=".client.EditProfileFragment">
 
 
     <TextView
