+++ app/src/main/res/layout/fragment_new_post.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/dark_purple"
-    tools:context=".NewPostFragment">
+    tools:context=".client.NewPostFragment">
 
     <TextView
         android:id="@+id/textView1"
