+++ app/src/main/res/layout/fragment_rate_post.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RatePostFragment">
+    tools:context=".client.RatePostFragment">
 
     <!-- TODO: Update blank fragment layout -->
